Add support for text/plain
This adds detection for the `text/plain` MIME type, with a file
extension of `txt`.

Detecting plain text is tricky, and it isn't very clearly spelled out in
any of the relevant RFCs (822, 2045, 2046). The best I can figure is "no
control characters" (00-1F), but even that isn't quite right: tab,
newline, and carriage return are control characters that often occur in
plain text.

---

While here, update the README with the list of all the `text` media
types.