Fix integer overflow in `msooxml` detection
Detection for MS-OOXML files will attempt to load a 32-bit address from
the user-provided payload, adding a static offset that may lead to
overflow if the original address is already near the max value for the
type used.

This commit has starting offset calculation be safe via use of
`checked_add`, returning `None` if the user-provided address is beyond
the bounds allowed.